Actually compiles in VS2012 now
(Probably won't in VS2010 though, since there's an emplace_back bug)

+++ src/Game/CollisionManager.cpp
@@ -98,7 +98,7 @@ Vector CollisionManager::CollisionClamp(const Entity& entity, const Vector& velo
     // Uh.
     static const std::function<float(TileData*&,Vector&)> distance = [](TileData*& tile, Vector& pos) -> float
     {
-        return (Vector(tile->x, tile->y, 0) - pos).length();
+        return static_cast<float>((Vector(tile->x, tile->y, 0) - pos).length());
     };
 	CollisionObj ent = getObj(entity);
